Bug 1043741 - Use the step result code for step failure determination
Previously only the presence of error lines found by the log parser
would cause the step to be marked as failing in the log viewer - even
if the step had a result value of !success. With this patch we use the
step result, thereby also handling cases where the log parser didn't
find a match & avoiding parser false positives from acting as a red
herring, if the actual failure occurred in a different step.