Bug 1098512 - check box filters work again
This is more heavy refactoring to make the queryString the single
source of truth.  Still not quite done, but this simplifies a fair bit
of code.  I also changed the ``isClassified`` to not be just ``true``
and ``false`` but used ``classified`` and ``unclassified`` as the
values as boolean was confusing.