Bug 1131133 - fix regression on job updates
The job poller was using a timestamp generated on the client to retrieve the updates.
Under some circumstances (eg a significant db replication lag) the
client was missing some updates.
Now it's using the latest last_modified fetched.
If for some reason we miss an update we re-run the same request on the next iteration