Bug 1136232 - Handle sessionStorage being disabled
Some users set dom.storage.enabled to false to disable local storage for
privacy reasons. Let's handle this more gracefully. In the future when
we use local storage for more features, we'll want to break this out to
somewhere common and use a shim of some sorts, but for now this will do.