Bug 1139204 - Don't reload UI every time we change state
Previously we were relying on URL routing to propagate changes when
a user added/removed/showed/hid a series (or changed the time interval).
This was pretty slow and annoying. Instead, just update the URL and
change just change the thing we need to change. Note that to make
this work, we have to import an in-development copy on angular-ui-router
with the deferIntercept function.