Bug 1153834 - Move job description tooltips to job details panel
Instead of displaying the long job description in the tooltip for every
job listed in the main jobs list, instead add it as the title text for
the job name in the job details panel. This avoids bloating the DOM with
thousands of longer strings.

We no longer check for the placeholder string "fill me", but I don't
think this matters, since we shouldn't be using that string in my
opinion anyway (vs leaving it blank) and to me it's more consistent to
always show a tooltip, rather than have the user wonder where it has
gone.