Bug 1163496 - Don't check if completed jobs are coalesced
With this change, we return early with the job result, if the job was
completed. This not only optimises for the common case, but also means
we avoid issues when job_coalesced_to_guid has been accidentally set for
completed jobs (eg due to the initial landing of bug 1093743, which has
since been fixed, but some old jobs are still in the DB).